+++ examples/passport/wallets-with-nextjs/src/app/connect-with-wagmi/config.ts
@@ -1,0 +1,12 @@
+import { createConfig, http } from "wagmi";
+import { immutableZkEvmTestnet } from "wagmi/chains";
+import { injected } from "wagmi/connectors";
+
+// create the Wagmi config for Immutable zkEVM Testnet
+export const config = createConfig({
+    chains: [immutableZkEvmTestnet],
+    connectors: [injected()],
+    transports: {
+      [immutableZkEvmTestnet.id]: http(),
+    },
+  });
