+++ examples/passport/wallets-with-nextjs/src/app/page.tsx
@@ -1,5 +1,3 @@
-'use client';
- 
 export default function Home() {
   
   return (<>
